# torus_mk2. performance. Convert function 'sign' in formulas as inline function