Sverchok master, Windows 11, Blender 3.6.1

![image](https://github.com/nortikin/sverchok/assets/14288520/90caa7bc-16c0-4c30-85e8-24848bc3d703)

235ms no optimisation

1 - 1x (111ms-104ms) (not usign sign function before and after optimization)
0 - 1.78x (212ms-119ms)
any - 1.65x (246ms-149ms)

with using of viewer node:
![image](https://github.com/nortikin/sverchok/assets/14288520/37a88b43-b423-41e7-a397-966d444f1412)


So, optimization is 1.5x-2.3x if do 'sign' function as inline call.