@satabol 2023-07-28
Append cache for edges and faces for some geometry params. It can speed up some cases.